requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cors==1.0.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.2.0
tweepy==4.8.0
tzdata==2022.1
urllib3==1.26.9
zope.event==4.5.0
zope.interface==5.4.0
